+++ src/index.ts
@@ -7,3 +7,5 @@ export {component, field, system} from './decorators';
 export {Query} from './query';
 export type {ComponentType} from './component';
 export type {Frame, SystemGroup} from './schedule';
+export {co, CanceledError} from './coroutines';
+export type {Coroutine, CurrentCoroutine} from './coroutines';